python-flask-qr-gen-v2/requirements.txt
Flask
qrcode[pil]
Pillow
